+++ app_modules/site/components/page/anchor/index.jsx
@@ -18,6 +18,7 @@ import { logCTAEvent } from 'app_modules/site/util/analytics';
 import componentUtil, { prefix as pf } from 'app_modules/ui/util/component';
 import navigation from 'app_modules/site/navigation/navigation';
 import { getActiveNavItems } from 'app_modules/site/navigation/navigation-utils';
+import Img from 'app_modules/ui/img';
 
 class Anchor extends React.Component {
 
@@ -27,7 +28,7 @@ class Anchor extends React.Component {
         <div className={pf('site-masthead-title col has-flexi-truncate align-middle')}>
           <div className={pf('media media--center media--responsive')}>
             <span className={pf('media__figure')}>
-              <img src={`/assets/images/header-${this.rootNavName()}.svg`} alt="" />
+              <Img src={`/assets/images/header-${this.rootNavName()}.svg`} alt="" />
             </span>
             <div className={pf('media__body')}>
               {this.renderBreadcrumbs()}