+++ app_modules/site/navigation/history.js
@@ -1,0 +1,6 @@
+import { createHistory, createMemoryHistory } from 'history';
+
+let history = createMemoryHistory();
+
+export const getHistory = () => history;
+export const setHistory = newHistory => history = newHistory;