+++ ui/components/feeds/flavors/comment-with-attachments/index.react.example.inline.jsx
@@ -14,13 +14,14 @@ const ButtonIcon = require('ui/components/buttons/flavors/icon/index.react');
 const SvgIcon = require('app_modules/ui/svg-icon');
 const StatefulClass = require('ui/components/lib/stateful.react');
 const MediaObject = require('ui/components/media-objects/index.react');
-const componentUtil = require('app_modules/ui/util/component');
-const pf = componentUtil.prefix;
+
+import Img from 'app_modules/ui/img';
+import { prefix as pf } from 'app_modules/ui/util/component';
 
 const image = (
   <div className={pf('avatar avatar--circle avatar--medium')}>
     <a href="javascript:void(0)" title="Jason Rodgers">
-      <img src={`/assets/images/avatar1.jpg`} alt='Jason Rodgers'/>
+      <Img src={`/assets/images/avatar1.jpg`} alt='Jason Rodgers'/>
     </a>
   </div>
 );
@@ -52,7 +53,7 @@ module.exports = (
     <div className={pf('attachments')}>
       <figure className={pf('attachments__item')}>
         <a href="javascript:void(0)" title="Marketing Plan - Rollout Q1 2016">
-          <img src={`/assets/images/feed/feed-image-example.png`} alt='Marketing Plan - Rollout Q1 2016'/>
+          <Img src={`/assets/images/feed/feed-image-example.png`} alt='Marketing Plan - Rollout Q1 2016'/>
         </a>
       </figure>
     </div>